requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-core
azure-cosmos
openai==0.28
numpy
requests
pandas
azure-storage-blob 
azure-identity
smart_open
tenacity
redis
tiktoken
azure-storage-file-share
python-dotenv
azure-search-documents>=11.4.0b3
azure-ai-formrecognizer
beautifulsoup4
lxml
azure-ai-textanalytics
langchain==0.0.139
flask
flask-socketio
openpyxl
